requirements.txt
flask
python-dotenv
supabase
gunicorn
